Try single placeholder for Navigation block
